refactor(web): split IntegrationSidebar into focused components
The sidebar component had a cognitive complexity of 60 (max 25). Decompose it
into a presentational shell + focused pieces, behavior unchanged:

- useIntegrationOwnership — shared own/forked derivation (was duplicated inline).
- IntegrationHeaderChips — connection + creator attribution chips.
- IntegrationActions — owns all connect/disconnect/publish/delete state,
  handlers, confirmation dialogs, and the bearer-token modal.
- NotConnectedActions / ConnectedActions — the two button-group layouts.

IntegrationSidebar is now a slim shell (header + tools + related workflows).
All five files are under the complexity threshold; type-check + lint pass.
